Rename _editing.py → utils.py
