+++ spd/editing/component_trainer.py
@@ -33,7 +33,7 @@ import torch.nn.functional as F
 from jaxtyping import Float, Int
 from torch import Tensor
 
-from spd.editing._editing import parse_component_key
+from spd.editing.utils import parse_component_key
 from spd.models.component_model import ComponentModel
