+++ spd/editing/utils.py
